Finished a pretty dope solar system

+++ engine/Engine/Apecs/Types.hs
@@ -10,20 +10,20 @@ import qualified Graphics.Rendering.OpenGL as GL
 import Language.Haskell.TH.Syntax
 import Engine.Vector (rotationMatrix)
 
-newtype Pos = Pos (V3 Float) deriving Show
-newtype Pos2D = Pos2D (V2 Float) deriving Show
-data Rot = RotEuler (V3 Float) 
-         | RotQuat (Quaternion Float)
-         | RotMat (M33 Float) 
+newtype Pos = Pos (V3 Double) deriving Show
+newtype Pos2D = Pos2D (V2 Double) deriving Show
+data Rot = RotEuler (V3 Double) 
+         | RotQuat (Quaternion Double)
+         | RotMat (M33 Double) 
          deriving Show
 
 getRotMat (RotEuler e) = rotationMatrix e
 getRotMat (RotQuat q) = fromQuaternion q
 getRotMat (RotMat m) = m
 
-newtype Scale = Scale (V3 Float) deriving Show
-newtype Size2D = Size2D (V2 Float) deriving Show
-data CamRot = CamRot Float Float deriving Show
+newtype Scale = Scale (V3 Double) deriving Show
+newtype Size2D = Size2D (V2 Double) deriving Show
+data CamRot = CamRot Double Double deriving Show
 data Hidden = Hidden deriving Show
 
 newtype Texture = Texture GL.TextureObject deriving Show
@@ -40,7 +40,7 @@ instance Component TerminalInput where type Storage TerminalInput = Unique Termi
 newtype Command = Command (Maybe Entity) deriving Show -- thats the previous command
 
 
-data TextBox = TextBox String Entity (V4 Float) Float deriving Show
+data TextBox = TextBox String Entity (V4 Double) Double deriving Show
 
 data FormControl = FormControl deriving Show -- bool is focus
 data Focus = Focus deriving Show
